test(tools): move cli test from main.rs to tools/cli.rs
Close #5
