+++ src/main.rs
@@ -1,9 +1,9 @@
+use parking_lot::Mutex;
 use std::collections::HashMap;
 use std::io::{Read, Write};
 use std::net::{SocketAddr, TcpListener, TcpStream, ToSocketAddrs};
 use std::process::id;
 use std::sync::Arc;
-use parking_lot::Mutex;
 
 use clap::Parser;
 use tokio::time::Duration;
@@ -30,7 +30,6 @@ use crate::session::session::Session;
 
 #[tokio::main]
 async fn main() {
-    
     // 启动参数解析
     let cli = crate::tools::cli::Cli::parse();
 
@@ -95,7 +94,9 @@ async fn main() {
     let arc_rdb_count = Arc::new(Mutex::new(RdbCount::new()));
     let arc_rdb_scheduler = Arc::new(Mutex::new(RdbScheduler::new(rdb)));
     if let Some(save_interval) = &rudis_config.save {
-        arc_rdb_scheduler.lock().execute(save_interval.clone(), arc_rdb_count.clone());
+        arc_rdb_scheduler
+            .lock()
+            .execute(save_interval.clone(), arc_rdb_count.clone());
     }
 
     for stream in listener.incoming() {
@@ -114,7 +115,8 @@ async fn main() {
                         sessions_clone,
                         rdb_count_clone,
                         aof_clone,
-                    ).await;
+                    )
+                    .await;
                 });
             }
             Err(e) => {
@@ -183,7 +185,6 @@ async fn connection(
                 read_size += size;
 
                 if size < 512 {
-                    
                     /*
                      * 解析命令
                      *
@@ -226,8 +227,7 @@ async fn connection(
                      * 否则响应 PONG 内容。
                      */
                     let uppercase_command = command.to_uppercase();
-                    if let Some(strategy) = command_strategies.get(uppercase_command.as_str())
-                    {
+                    if let Some(strategy) = command_strategies.get(uppercase_command.as_str()) {
                         // 执行命令
                         strategy.execute(
                             Some(&mut stream),
@@ -301,91 +301,4 @@ fn println_banner(port: u16) {
         id()
     );
     println!("{}", pattern);
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::db::db_config::RudisConfig;
-    use crate::tools::cli;
-    use clap::Parser;
-    #[test]
-    fn test_config_file() {
-        let port = 42800;
-        let config_file_path = "./release/linux/rudis-server.properties";
-        let arg_string = format!(
-            "rudis-server \
-            -p {} \
-            --config {}",
-            port, config_file_path
-        );
-        let args: Vec<&str> = arg_string.split(' ').collect();
-        let cli = cli::Cli::parse_from(args);
-        let config: RudisConfig = cli.into();
-        assert_eq!(config.port, port);
-        assert_eq!(config.maxclients, 1000);
-        assert_eq!(config.password, None);
-        assert_eq!(config.save.unwrap().iter().map(|x| format!("{}/{}",x.0,x.1)).collect::<Vec<String>>().join(" "), "60/1 20/2");
-    }
-    #[test]
-    fn test_cli() {
-        let bind = "192.168.1.2";
-        let port = 6379;
-        let password = "123456";
-        let databases = 1;
-        let dbfilename = "123.rdb";
-        let appendfilename = "321.aof";
-        let appendonly = "false";
-        let hz = 2;
-        let appendfsync = "asd";
-        let maxclients = 100;
-        let dir = "/home/rudis";
-        let save_1 = "60/1";
-        let save_2 = "20/2";
-        let save = "60/1 20/2";
-        let arg_string = format!(
-            "rudis-server \
-            --bind {} \
-            -p {} \
-            --password {} \
-            --databases {} \
-            --dbfilename {} \
-            --appendfilename {} \
-            --hz {} \
-            --appendfsync {} \
-            --maxclients {} \
-            --dir {} \
-            --save {} \
-            --save {} \
-            --appendonly {}",
-            bind,
-            port,
-            password,
-            databases,
-            dbfilename,
-            appendfilename,
-            hz,
-            appendfsync,
-            maxclients,
-            dir,
-            save_1,
-            save_2,
-            appendonly
-        );
-
-        let args: Vec<&str> = arg_string.split(' ').collect();
-        let cli = cli::Cli::parse_from(args);
-        let config: RudisConfig = cli.into();
-        assert_eq!(config.bind, bind.to_string());
-        assert_eq!(config.port, port);
-        assert_eq!(config.password, Some(password.to_string()));
-        assert_eq!(config.databases, databases);
-        assert_eq!(config.dbfilename, Some(dbfilename.to_string()));
-        assert_eq!(config.appendfilename, Some(appendfilename.to_string()));
-        assert_eq!(config.appendonly.to_string(), appendonly);
-        assert_eq!(config.appendfsync, Some(appendfsync.to_string()));
-        assert_eq!(config.maxclients, maxclients);
-        assert_eq!(config.dir, dir.to_string());
-        assert!(config.save.is_some());
-        assert_eq!(config.save.unwrap().iter().map(|x| format!("{}/{}",x.0,x.1)).collect::<Vec<String>>().join(" "), save);
-    }
 }
